requirements.txt
streamlit
faiss-cpu
langchain
langchain-community
langchain-openai
openai
psycopg2-binary
python-dotenv
sqlalchemy
matplotlib
numpy
seaborn